backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
pyotp==2.9.0
SQLAlchemy==2.0.23
googlemaps==4.10.0
requests==2.31.0
python-socketio==5.10.0
eventlet==0.33.3
flask-restx==1.3.0
Flask-Mail
